feat: add DropdownSection component and integrate it into SelectWithTagGroup and App components
