+++ src/App.tsx
@@ -27,6 +27,7 @@ import { ToggleSectionField } from "./field/ToggleSectionField";
 import { useAppForm } from "./form";
 import { Link } from "./link/Link";
 import { DropdownItem } from "./listbox/DropdownItem";
+import { DropdownSection } from "./listbox/DropdownSection";
 import { ListBox } from "./listbox/ListBox";
 import { ListBoxItem } from "./listbox/ListBoxItem";
 import { VirtualizedListBox } from "./listbox/VirtualizedListBox";
@@ -172,6 +173,32 @@ export const App = () => {
           </DropdownItem>
         )}
       </SelectWithTagGroup>
+
+      <Select label="Select an animal with sections" size="sm">
+        <ListBox>
+          <DropdownSection title="Mammals" items={animals.filter((_, i) => i < 10)}>
+            {(item) => (
+              <DropdownItem key={item.id} id={String(item.id)} textValue={item.name}>
+                {item.name}
+              </DropdownItem>
+            )}
+          </DropdownSection>
+          <DropdownSection title="Birds" items={animals.filter((_, i) => i >= 10 && i < 15)}>
+            {(item) => (
+              <DropdownItem key={item.id} id={String(item.id)} textValue={item.name}>
+                {item.name}
+              </DropdownItem>
+            )}
+          </DropdownSection>
+          <DropdownSection title="Other Animals" items={animals.filter((_, i) => i >= 15)}>
+            {(item) => (
+              <DropdownItem key={item.id} id={String(item.id)} textValue={item.name}>
+                {item.name}
+              </DropdownItem>
+            )}
+          </DropdownSection>
+        </ListBox>
+      </Select>
 
       <Select label="Select an animal" size="sm">
         <ListBox items={animals}>
